Additional CI speed-ups
Adding a cache for up-cpp build artifacts cuts another 30 seconds out of
the best-case build times (i.e. when cache is valid).

The same technique of checking the HEAD hash for the upstream repo and
providing that as the cache key is used here. Whenever HEAD is changed
in up-cpp (typically a new PR is merged), the cache is invalidated and
up-cpp is rebuilt.

Also cuts ~10 seconds out by reusing the caches we already have in the
up-client-zenoh-cpp build step. Caches load faster than uploading and
downloading additional artifacts.