Fix issues with caches in PRs changing ci.yml (#35)
GitHub will fall back to the cache from the target branch of a PR if it
fails to find a cache for the source branch of a PR. This causes issues
if the PR is changing the ci.yml file (see #32) without one of the
dependency repos changing (up-cpp and zenoh-c).

This change adds the SHA of the workflow file to the cache key so that
changes to ci.yml will invalidate the caches.

While this reduces the value of the caches when iterating on ci.yml
changes, a developer working in ci.yml would have the ability to remove
this temporarily if needed.